api/LamdaLayer/requirements.txt
aws-lambda-powertools
pycryptodome
pyjwt